moved batch_delete route to source_app/views.py
