+++ securedrop/source.py
@@ -1,18 +1,14 @@
 # -*- coding: utf-8 -*-
 from cStringIO import StringIO
 from flask import (request, render_template, session, redirect, url_for,
-                   flash, g, send_file, Markup, make_response)
+                   flash, send_file, Markup, make_response)
 
 import config
 import json
 import version
 import crypto_util
 from flask_babel import gettext
-from rm import srm
-import store
-from db import db_session
 from source_app import create_app
-from source_app.decorators import login_required
 from source_app.utils import logged_in, valid_codename
 
 import logging
@@ -21,22 +17,6 @@ import logging
 log = logging.getLogger('source')
 
 app = create_app()
-
-
-@app.route('/delete-all', methods=('POST',))
-@login_required
-def batch_delete():
-    replies = g.source.replies
-    if len(replies) == 0:
-        app.logger.error("Found no replies when at least one was expected")
-        return redirect(url_for('main.lookup'))
-    for reply in replies:
-        srm(store.path(g.filesystem_id, reply.filename))
-        db_session.delete(reply)
-    db_session.commit()
-
-    flash(gettext("All replies have been deleted"), "notification")
-    return redirect(url_for('main.lookup'))
 
 
 @app.route('/login', methods=('GET', 'POST'))
