+++ securedrop/source_templates/lookup.html
@@ -77,7 +77,7 @@
         <div class="clearfix"></div>
       </div>
     {% endfor %}
-    <form id="delete-all" method="post" action="/delete-all">
+    <form id="delete-all" method="post" action="{{ url_for('main.batch_delete') }}">
       <a class="sd-button btn" href="#delete-all-confirm">{{ gettext('DELETE ALL REPLIES') }}</a>
       <input name="csrf_token" type="hidden" value="{{ csrf_token() }}">
       <div id="delete-all-confirm" class="hidden-prompt">
